requirements.txt
opencv-python==4.5.5.64
PyYAML==5.3.1
tensorflow==2.12.0
boto3==1.26.95
pandas
numpy
pytesseract
python-dotenv
